% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{animate_local_search}
\alias{animate_local_search}
\title{Generate animation of the local search clustering approach}
\usage{
animate_local_search(clust, filename = "animation.gif")
}
\arguments{
\item{clust}{A list returned by clustering function with cluster_method = "local_search"}

\item{filename}{The filename of the gif to be saved}
}
\value{
Nothing, saves a .gif file in the current workdir
}
\description{
Generate animation of the local search clustering approach
}
