% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_eval_cpp}
\alias{cluster_eval_cpp}
\title{Evaluate a cluster according to average distance divided by total profit}
\usage{
cluster_eval_cpp(zone, inst)
}
\arguments{
\item{zone}{A numeric vector of the id's included in the zone}

\item{inst}{A data.matrix of the instance contatining the columns id, x, y, score. Row number should coincide with point id.}
}
\description{
Evaluate a cluster according to average distance divided by total profit
}
