% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Generate clusters based on an instance}
\usage{
clustering(inst, k, cluster_method = "pam")
}
\arguments{
\item{inst}{An object returned from the `instance` function}

\item{k}{The number of clusters}

\item{cluster_method}{The method with which to perform the clustering}
}
\value{
A list ...
}
\description{
Given an instance, number of clusters and a clustering method the function assigns each node of the problem instance a zone.
}
