% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_point_segment.R
\name{distancePointSegment}
\alias{distancePointSegment}
\title{Calculate distance from point to line segment}
\usage{
distancePointSegment(px, py, x1, y1, x2, y2)
}
\arguments{
\item{px}{x-coordinate for point}

\item{py}{y-coordinate for point}

\item{x1}{x-coordinate for the first point from the line segement}

\item{y1}{y-coordinate for the first point from the line segement}

\item{x2}{x-coordinate for the second point from the line segement}

\item{y2}{y-coordinate for the second point from the line segement}
}
\value{
numeric value
}
\description{
Calculate distance from point to line segment
}
