% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{generate_information}
\alias{generate_information}
\title{Generate an information set}
\usage{
generate_information(inst, dst, r = 20)
}
\arguments{
\item{inst}{A list returned from the `instance` function}

\item{dst}{Shortest path distances from igraph}

\item{r}{The maximum radius information can travel}
}
\value{
a matrix of information
}
\description{
Given an instance the function will generate a an information matrix.
}
