% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{generate_variances}
\alias{generate_variances}
\title{Generate a set of variances for the stochastic profits}
\usage{
generate_variances(
  inst,
  distribution = "discrete_uniform",
  bounds = c(min = 10, max = 40)
)
}
\arguments{
\item{inst}{A list returned from the `instance` function}

\item{distribution}{A distribution to use, for only "discrete_uniform" is available}

\item{bounds}{Bounds to use for the distribution in the uniform case}
}
\value{
a tibble with `id` and `score_variance`
}
\description{
Given an instance the function will generate a variance number for each individual node.
}
