% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_route2.R
\name{initial_route2}
\alias{initial_route2}
\title{Generation of initial routes as part of pre-mission cluster - 2nd edition}
\usage{
initial_route2(p_inst, L, info, top_percentile = 0.5)
}
\arguments{
\item{p_inst}{A list returned from the `prepare_instance` function}

\item{L}{The range for each agent}

\item{info}{An information matrix returned by the `generate_information` function}
}
\value{
A route satisfying the range constraint, represented as an integer vector
}
\description{
This functions will the next node among all nodes relative to SDR as opposed to the first edition that sample only the first point.
}
