% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{instance}
\alias{instance}
\title{Instantiate a test instance}
\usage{
instance(path_to_file)
}
\arguments{
\item{path_to_file}{Path to the `.txt` file holding the instance data}
}
\value{
`instance` returns an object of \code{\link{class}} "`instance`".
  An object of class "`instance`" is a list containing:
  `points` a `tibble` containing coordinates and scores for each point.
  `n` an integer equal to the number of points in the instance.
}
\description{
`instance` is used to instantiate a test instance.
There are some test instances included in the package, both as raw data and as instance objects.
This function is used to convert the raw data to instance objects.
}
