% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_routes.R
\name{plot.initial_route}
\alias{plot.initial_route}
\title{Plot the initial route on the instance}
\usage{
\method{plot}{initial_route}(init_route, inst)
}
\arguments{
\item{init_route}{A list returned by the `initial_route` function}

\item{inst}{}
}
\value{

}
\description{
Plot the initial route on the instance
}
