% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_route2.R
\name{plot.initial_route2}
\alias{plot.initial_route2}
\title{Plot the initial route on the instance}
\usage{
\method{plot}{initial_route2}(init_route, p_inst)
}
\arguments{
\item{init_route}{A list returned by the `initial_route` function}

\item{inst}{}
}
\value{

}
\description{
Plot the initial route on the instance
}
