% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_instance.R
\name{plot.prepared_instance}
\alias{plot.prepared_instance}
\title{Plot method for a prepared_instance}
\usage{
\method{plot}{prepared_instance}(p_inst)
}
\arguments{
\item{p_inst}{A object returned from the `prepare_instance` function}
}
\value{
A ggplot2 object
}
\description{
Building on the plot method from the instance object, this function adds information about socre variance and whether an unexpected event occurs at a node.
}
