% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_routes.R
\name{plot.rb_clustering}
\alias{plot.rb_clustering}
\title{Plot the result from routing-based clustering}
\usage{
\method{plot}{rb_clustering}(rb_clust)
}
\arguments{
\item{rb_clust}{}
}
\value{

}
\description{
Plot the result from routing-based clustering
}
