% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{plot.routing}
\alias{plot.routing}
\title{Plot method for a routing object}
\usage{
\method{plot}{routing}(rout)
}
\arguments{
\item{rout}{A list returned from the `routing` function}
}
\value{
A ggplot object
}
\description{
Visualize the generated routes
}
