% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing2.R
\name{plot.starting_routes}
\alias{plot.starting_routes}
\title{Plot method for the starting routes}
\usage{
\method{plot}{starting_routes}(sr, inst)
}
\arguments{
\item{inst}{}
}
\value{

}
\description{
Plot method for the starting routes
}
