% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing2.R
\name{plot.updated_routes}
\alias{plot.updated_routes}
\title{Plot method for the updated routes}
\usage{
\method{plot}{updated_routes}(ur, inst)
}
\arguments{
\item{inst}{}
}
\description{
Plot method for the updated routes
}
