% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_routes.R
\name{plot.updated_routes2}
\alias{plot.updated_routes2}
\title{Plot method for the updated routes}
\usage{
\method{plot}{updated_routes2}(ur, inst)
}
\arguments{
\item{inst}{}
}
\description{
Plot method for the updated routes
}
