% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_instance.R
\name{prepare_instance}
\alias{prepare_instance}
\title{Prepare an instance for use with clustering and routing algorithms}
\usage{
prepare_instance(inst, variances, info)
}
\arguments{
\item{variances}{A list returned from the `generate_variances` function}

\item{info}{A list returned from the `generate_information` function}

\item{name}{The name of the instance to use, eligible names are those in `names(dz::test_instances)`}
}
\value{
A list, i.e. a "prepared" instance
}
\description{
Generates variances and information and joins the information on the original instance
}
