% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_routes.R
\name{rb_clustering}
\alias{rb_clustering}
\title{Routing based clustering}
\usage{
rb_clustering(
  p_inst,
  L,
  k,
  num_routes,
  info,
  top_percentile = 0.15,
  dispute_obj = "most_frequent",
  measure = "position-based",
  weigthed = T,
  shiny = F
)
}
\arguments{
\item{shiny}{}
}
\value{

}
\description{
Routing based clustering
}
