% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_routes.R
\name{resolve_disputes}
\alias{resolve_disputes}
\title{Resolve disputes so each node are only assigned one cluster}
\usage{
resolve_disputes(init_routes, cluster, obj = "most_frequent")
}
\arguments{
\item{init_routes}{a list of initial routes obtained from the `initial_route` function}

\item{cluster}{a vector of clusters obtained from the `route_clustering` function}

\item{obj}{what objective to use when solving the disputes, currently there are "most_frequent" and "highest_score"}
}
\value{
list containing the zones as well as the zones of only undisputed nodes
}
\description{
Resolve disputes so each node are only assigned one cluster
}
