% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{routing}
\alias{routing}
\title{Generate routes for a clustered instance}
\usage{
routing(clust, routing_method = "random")
}
\arguments{
\item{clust}{A list returned from the `clustering` function}

\item{routing_method}{A method from which to perform the routing}
}
\value{
A list
}
\description{
Given a clustered instance and a routing method the function will provide routes for the given instance.
}
