% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing2.R
\name{starting_routes}
\alias{starting_routes}
\title{Find the starting routes based on provided zones}
\usage{
starting_routes(inst, zones, L, L_budget = c(initial = 0.6, improve = 0.8))
}
\arguments{
\item{L}{}
}
\description{
Find the starting routes based on provided zones
}
