% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_instances}
\alias{test_instances}
\title{7 team orienteering problem test instances}
\format{
A list of `instance` objects
}
\usage{
test_instances
}
\description{
A list of 7 different test instances for the team orienteering problem of variying size.
They have been obtained from \url{https://www.mech.kuleuven.be/en/cib/op} under the team orienteering problem section.
}
\keyword{datasets}
