% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing2.R
\name{update_routes}
\alias{update_routes}
\title{Update the starting routes based on realized scores}
\usage{
update_routes(sr, L, variances, info)
}
\arguments{
\item{info}{}
}
\description{
Update the starting routes based on realized scores
}
