% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_routes.R
\name{update_routes2}
\alias{update_routes2}
\title{update routes based on the new information available during mission}
\usage{
update_routes2(p_inst, zones, L, k, sr, info)
}
\arguments{
\item{info}{}
}
\description{
update routes based on the new information available during mission
}
