% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Generate clusters based on an instance}
\usage{
clustering(
  inst,
  k,
  L,
  eps,
  variances,
  info,
  cluster_method = c("greedy", "local_search"),
  alpha = 1
)
}
\arguments{
\item{inst}{An object returned from the `instance` function}

\item{k}{The number of clusters}

\item{L}{The length constraint for each agent}

\item{eps}{A sensitivity parameter for the preprocessing of the instance with regards to excluding points that are too far away.}

\item{cluster_method}{The method with which to perform the clustering}

\item{alpha}{weight between mean distance/total profit and relevancy}
}
\value{
A list ...
}
\description{
Given an instance, number of clusters and a clustering method the function assigns each node of the problem instance a zone.
}
