% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_routes.R
\name{fix_connectivity}
\alias{fix_connectivity}
\title{Fix connectivity in zones}
\usage{
fix_connectivity(inst, zones)
}
\arguments{
\item{zones}{}
}
\value{

}
\description{
Fix connectivity in zones
}
