% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_routes.R
\name{initial_route}
\alias{initial_route}
\title{Generation of inital routes as part of the pre-mission clustering}
\usage{
initial_route(inst, L, r, variances)
}
\arguments{
\item{inst}{Instance object returned from the `instance` function}

\item{L}{The range for each agent (homogenous agents are assumed)}

\item{r}{The radius from a node that profits will be realized}

\item{variances}{A set of variances returned from the `generate_variances` function}
}
\value{
A route satisfying the range constraint, represented as an integer vector
}
\description{
Generation of inital routes as part of the pre-mission clustering
}
