% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot.clustering}
\alias{plot.clustering}
\title{Plot method for a clustering object}
\usage{
\method{plot}{clustering}(clust, delaunay = T)
}
\arguments{
\item{clust}{A list returned from the `clustering` function}
}
\value{
A ggplot object
}
\description{
Visualizes the zones obtained from the specific clustering algorithm.
}
