% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{plot.instance}
\alias{plot.instance}
\title{Plot method for instance object}
\usage{
\method{plot}{instance}(inst, delaunay = FALSE, voronoi = FALSE)
}
\arguments{
\item{inst}{Object of class `instance`}
}
\value{
A ggplot object
}
\description{
Plot method for instance object
}
