% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_routes.R
\name{rb_clustering}
\alias{rb_clustering}
\title{Routing based clustering}
\usage{
rb_clustering(inst, L, k, variances, info)
}
\arguments{
\item{info}{}
}
\value{

}
\description{
Routing based clustering
}
