% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{routing}
\alias{routing}
\title{Generate routes for a clustered instance}
\usage{
routing(clust, obj = "SDR", L = 500, variances)
}
\arguments{
\item{clust}{A list returned from the `clustering` function}

\item{obj}{The objective to use for the routing}

\item{L}{The range constraint induced on the agents}
}
\value{
A list
}
\description{
Given a clustered instance and a routing method the function will provide routes for the given instance.
}
