% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing2.R
\name{starting_routes}
\alias{starting_routes}
\title{Find the starting routes based on provided zones}
\usage{
starting_routes(inst, zones, L)
}
\arguments{
\item{L}{}
}
\value{

}
\description{
Find the starting routes based on provided zones
}
